compute average score

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -8,7 +8,7 @@ CREATE PROCEDURE `ComputeAverageScoreForUser`(
     IN `user_id` INT
 )
 BEGIN
-    DECLARE avg_score DECIMAL(10, 2); -- Adjust precision and scale based on your needs
+    DECLARE avg_score FLOAT; -- Adjust precision and scale based on your needs
     
     -- Compute average score for the given user_id
     SELECT AVG(score) INTO avg_score FROM corrections WHERE user_id = user_id;
